+++ common-core/src/main/java/cn/featherfly/common/structure/ChainMap.java
@@ -14,7 +14,28 @@ public interface ChainMap<K, V> extends Map<K, V> {
     /**
      * 带链式调用的put方法. {@link java.util.Map#put(Object, Object)}.
      *
-     * @param key   指定值将要关联的键。
+     * @param key 指定值将要关联的键。
+     * @param value 指定键将要关联的值。
+     * @return 当前MAP
+     */
+    default ChainMap<K, V> set(K key, V value) {
+        return putChain(key, value);
+    }
+
+    /**
+     * 带链式调用的putAll方法. {@link java.util.Map#putAll(Map)}.
+     *
+     * @param map 要在此映射中存储的映射关系
+     * @return 当前MAP
+     */
+    default ChainMap<K, V> set(Map<? extends K, ? extends V> map) {
+        return putAllChain(map);
+    }
+
+    /**
+     * 带链式调用的put方法. {@link java.util.Map#put(Object, Object)}.
+     *
+     * @param key 指定值将要关联的键。
      * @param value 指定键将要关联的值。
      * @return 当前MAP
      */
@@ -23,8 +44,8 @@ public interface ChainMap<K, V> extends Map<K, V> {
     /**
      * 带链式调用的putAll方法. {@link java.util.Map#putAll(Map)}.
      *
-     * @param m 要在此映射中存储的映射关系
+     * @param map 要在此映射中存储的映射关系
      * @return 当前MAP
      */
-    ChainMap<K, V> putAllChain(Map<? extends K, ? extends V> m);
+    ChainMap<K, V> putAllChain(Map<? extends K, ? extends V> map);
 }